Fixed Issues and successfully connected to database
