+++ EventDriven.Project.Enrollment/EventDriven.Project.Enrollment.csproj
@@ -1,0 +1,15 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net8.0-windows</TargetFramework>
+    <Nullable>enable</Nullable>
+    <UseWindowsForms>true</UseWindowsForms>
+    <ImplicitUsings>enable</ImplicitUsings>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\EventDriven.Project.Logic\EventDriven.Project.Logic.csproj" />
+    <ProjectReference Include="..\EventDriven.Project.Model\EventDriven.Project.Model.csproj" />
+  </ItemGroup>  
+
+</Project>
