+++ EventDriven.Project.Enrollment/EventDriven.Project.Enrollment.csproj.user
@@ -1,0 +1,2 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003" />
