--- EventDriven.Project.Enrollment/EventDriven.Project.UI.csproj
@@ -1,20 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<PropertyGroup>
-		<OutputType>WinExe</OutputType>
-		<TargetFramework>net8.0-windows</TargetFramework>
-		<Nullable>enable</Nullable>
-		<UseWindowsForms>true</UseWindowsForms>
-		<ImplicitUsings>enable</ImplicitUsings>
-	</PropertyGroup>
-
-	<PropertyGroup>
-		<StartupObject>EventDriven.Project.UI.Program</StartupObject>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<ProjectReference Include="..\EventDriven.Project.Logic\EventDriven.Project.Logic.csproj" />
-		<ProjectReference Include="..\EventDriven.Project.Model\EventDriven.Project.Model.csproj" />
-	</ItemGroup>
-
-</Project>
