--- EventDriven.Project.Enrollment/EventDriven.Project.UI.csproj.user
@@ -1,11 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <ItemGroup>
-    <Compile Update="Form1.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Update="TempLogin.cs">
-      <SubType>Form</SubType>
-    </Compile>
-  </ItemGroup>
-</Project>
