+++ EventDriven.Project.UI/Program.cs
@@ -1,0 +1,29 @@
+namespace EventDriven.Project.UI
+{
+    internal static class Program
+    {
+        /// <summary>
+        ///  The main entry point for the application.
+        /// </summary>
+        [STAThread]
+        static void Main()
+        {
+            // To customize application configuration such as set high DPI settings or default font,
+            // see https://aka.ms/applicationconfiguration.
+
+            ApplicationConfiguration.Initialize();
+            using (TempLogin loginForm = new TempLogin())
+            {
+                if (loginForm.ShowDialog() == DialogResult.OK)
+                {
+                    Application.Run(new Form1());
+                }
+                else
+                {
+                    Application.Exit();
+
+                }
+            }
+        }
+    }
+}
